# [PERF] mrp: fix redundant phantom BOM calculations