@robodoo 2025-09-19
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/227894.png)](https://mergebot.odoo.com/odoo/odoo/pull/227894)